requirements.txt
streamlit
SQLAlchemy
psycopg2-binary
ucimlrepo 